Also, cast index to `int` - don't want any php surprises appearing somewhere

+++ src/php/Board.php
@@ -148,7 +148,7 @@ class Board
 
         $keys = array_filter($this->flat_board, $callback);
         foreach (array_keys($keys) as $index) {
-            $moves[] = [floor($index / $this->size), $index % $this->size];
+            $moves[] = [(int) floor($index / $this->size), $index % $this->size];
         }
         return $moves;
     }
